+++ extensions/gitlab/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Fix crash] - 2022-12-07
 
 - Fix crash which happens when cache is corrupted
+- Add `Clear Local Extension Cache` action to most list items to be able to reset the local cache
 
 ## [Todos Menu Bar] - 2022-12-06